conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.9.0=py38h32f6830_1
blas=1.0=mkl
c-ares=1.15.0=h01d97ff_1001
ca-certificates=2020.4.5.1=hecc5488_0
certifi=2020.4.5.1=py38h32f6830_0
cycler=0.10.0=py_2
freetype=2.10.2=h8da9a1a_0
grpcio=1.23.0=py38hb05251e_1
importlib-metadata=1.6.0=py38h32f6830_0
jpeg=9c=h1de35cc_1001
kiwisolver=1.2.0=py38ha0d09dd_0
libcxx=10.0.0=h1af66ff_2
libffi=3.2.1=h4a8c4bd_1007
libgfortran=3.0.1=0
libpng=1.6.37=hbbe82c9_1
libprotobuf=3.11.4=hd174df1_0
libtiff=4.1.0=h2ae36a8_6
libwebp-base=1.1.0=h0b31af3_3
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.2=h4a8c4bd_1
markdown=3.2.2=py_0
matplotlib=3.2.1=0
matplotlib-base=3.2.1=py38h1300a51_0
mkl=2019.5=281
mkl-service=2.3.0=py38h0b31af3_0
mkl_fft=1.1.0=py38h3b54f70_1
mkl_random=1.1.0=py38h4a8c4bd_0
ncurses=6.1=h0a44026_1002
ninja=1.10.0=ha1b3eb9_0
numpy=1.18.1=py38h7241aed_0
numpy-base=1.18.1=py38h6575580_1
olefile=0.46=py_0
openssl=1.1.1g=h0b31af3_0
pillow=7.1.2=py38h83dc5e5_0
pip=20.1=pyh9f0ad1d_0
protobuf=3.11.4=py38hc84c608_1
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.8.2=hd5f0129_7_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.5.0=py3.8_0
readline=8.0=hcfe32e1_0
scipy=1.4.1=py38h44e99c9_0
setuptools=46.3.1=py38h32f6830_0
six=1.14.0=py_1
sqlite=3.30.1=h93121df_0
tensorboard=1.15.0=py38_0
tk=8.6.10=hbbe82c9_0
torchvision=0.6.0=py38_cpu
tornado=6.0.4=py38h64e0658_1
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.34.2=py_1
xz=5.2.5=h0b31af3_0
zipp=3.1.0=py_0
zlib=1.2.11=h0b31af3_1006
zstd=1.4.4=h4b3e974_3
